ref(calendar): move meeting url parser to common utils
To be used by other features for validating if an
adhoc meeting is on a whitelisted domain.
